backend/requirements.txt
# Web Framework & Server
fastapi>=0.115.0
uvicorn>=0.34.0
bcrypt
# Explicitly pin websockets if uvicorn throws warnings
websockets>=13.0

# Database Drivers
# MongoDB & Vector DB
pymongo
qdrant-client

# SQL Drivers (PostgreSQL via SQLAlchemy)
sqlalchemy
psycopg2-binary

# Security & Utilities
pydantic>=2.0.0
python-dotenv
pip-system-certs

# AI & Agent Frameworks
google-genai
google-api-core

# Parlant 3.2 Update Configuration
parlant[gemini,ollama]==3.2.0
parlant-client==3.2.0
# Parlant 3.1/3.2 requires tokenizers>=0.21
tokenizers>=0.21
# Transformers must be >=4.49.0 to support tokenizers 0.21
transformers>=4.49.0
# Required for local embeddings
sentence-transformers

# Data Processing & File Handling
pandas
numpy
openpyxl
xlrd
hf_xet
watchdog

# Access to MinIO
minio
aiohttp